requirements.txt
streamlit
plotly
geopandas
shapely
geopy
pytesseract
opencv-python-headless
PyMuPDF
pdf2image
transformer